added Profiler to the codebase, added some methods to profiler scope, lil Time change, miliseconds now uses float instead of uint32_t, added HoveredRect to Helpers

+++ include/BeatEngine/System/Time.h
@@ -12,7 +12,7 @@ public:
     constexpr Time(const std::chrono::duration<Rep, Period>& duration);
 
     constexpr float AsSeconds() const;
-    constexpr int32_t AsMilliseconds() const;
+    constexpr float AsMilliseconds() const;
     constexpr int64_t AsMicroseconds() const;
     constexpr std::chrono::time_point<std::chrono::system_clock> AsTimePoint() const;
     constexpr time_t AsTimeT() const;
